requirements-test.txt
-e .
pytest==6.0.1
pytest-aiohttp==0.3.0
pytest-asyncio==0.14.0
mypy==0.782
flake8==3.8.3
flake8-isort==4.0.0
black==19.10b0
isort==5.4.2
